fix: complete FP8+Eagle3 dtype compatibility for MoE targets
Fix multiple dtype mismatches that prevented Eagle3 from running with
native FP8 target models (required --dtype bfloat16 workaround):

1. logits_processor.py: Add bfloat16 cast in is_full() aux_hidden_states
   path (matching existing is_last() fix)
2. eagle_worker.py: Use bfloat16 for idle input hidden_states instead of
   target model dtype (float16/float32 with FP8)
3. eagle_info.py: Same idle input dtype fix in verify output paths
4. llama.py: Cast shared embedding and lm_head to bfloat16 in
   set_embed() and set_embed_and_head() (FP8 targets may have
   float16/float32 embeddings that mismatch draft model bfloat16)

Tested: MiniMax-M2.5 (229B MoE) now starts and serves with Eagle3
in native FP8 mode without CUDA graph capture errors.

+++ python/sglang/srt/speculative/eagle_worker.py
@@ -497,7 +497,7 @@ class EAGLEWorker(TpModelWorker):
         batch.spec_info = EagleDraftInput.create_idle_input(
             device=self.device,
             hidden_size=self.model_config.hidden_size,
-            dtype=self.model_config.dtype,
+            dtype=torch.bfloat16,
             topk=self.topk,
             capture_hidden_mode=CaptureHiddenMode.LAST,
         )
@@ -936,7 +936,7 @@ class EAGLEWorker(TpModelWorker):
             batch.spec_info = EagleDraftInput.create_idle_input(
                 device=self.device,
                 hidden_size=hidden_size,
-                dtype=self.model_config.dtype,
+                dtype=torch.bfloat16,
                 topk=self.topk,
                 capture_hidden_mode=CaptureHiddenMode.LAST,
             )
